+++ src/neuroconv/datainterfaces/behavior/fictrac/fictracdatainterface.py
@@ -156,7 +156,8 @@ class FicTracDataInterface(BaseTemporalAlignmentInterface):
         file_path : a string or a path
             Path to the .dat file (the output of fictrac)
         radius : float, optional
-            The radius of the ball in meters. If provided the data will be converted to meters and stored as such.
+            The radius of the ball in meters. If provided the radius is stored as a conversion factor
+            and the units are set to meters. If not provided the units are set to radians.
         configuration_file_path : a string or a path, optional
             Path to the .txt file with the configuration metadata. Usually called config.txt
         verbose : bool, default: True
@@ -235,7 +236,7 @@ class FicTracDataInterface(BaseTemporalAlignmentInterface):
             column_in_dat_file = data_dict["column_in_dat_file"]
             data = fictrac_data_df[column_in_dat_file].to_numpy()
             if self.radius is not None:
-                data = data * self.radius
+                spatial_series_kwargs["conversion"] = self.radius
                 units = "meters"
             else:
                 units = "radians"
@@ -252,8 +253,8 @@ class FicTracDataInterface(BaseTemporalAlignmentInterface):
             spatial_series = SpatialSeries(**spatial_series_kwargs)
             position_container.add_spatial_series(spatial_series)
 
-        # Add the compass direction container to the processing module
-        processing_module = get_module(nwbfile=nwbfile, name="Behavior")
+        # Add the container to the processing module
+        processing_module = get_module(nwbfile=nwbfile, name="behavior")
         processing_module.add_data_interface(position_container)
 
     def get_original_timestamps(self):